Separate test infrastruture from tests.
