+++ numba/cuda/tests/nocuda/__init__.py
@@ -1,9 +1,6 @@
-"""
-This subpackage contains CUDA tests that can be executed without a CUDA driver.
-"""
-
-from numba.tests import SerialSuite
+from numba.testing import SerialSuite
+from numba.testing import load_testsuite
+import os
 
 def load_tests(loader, tests, pattern):
-    suite = loader.discover("numba.cuda.tests.nocuda")
-    return SerialSuite(suite)
+    return SerialSuite(load_testsuite(loader, os.path.dirname(__file__)))
